+++ python/hugr_qir/guppy_to_qir.py
@@ -2,10 +2,16 @@ from collections.abc import Callable
 
 from guppylang.defs import GuppyFunctionDefinition
 
+from ._hugr_qir import max_loop_unroll_default
 from .hugr_to_qir import to_qir_bytes, to_qir_str
 
 
-def guppy_to_qir_bytes(entrypoint: Callable, *, validate_qir: bool = True) -> bytes:
+def guppy_to_qir_bytes(
+    entrypoint: Callable,
+    *,
+    validate_qir: bool = True,
+    max_loop_unroll: int = max_loop_unroll_default(),
+) -> bytes:
     """
     Converts guppy entrypoint to qir bytes
 
@@ -13,6 +19,7 @@ def guppy_to_qir_bytes(entrypoint: Callable, *, validate_qir: bool = True) -> by
     :type entrypoint: GuppyFunctionDefinition
     :param validate_qir: Whether to validate the created QIR
     :type validate_qir: bool
+    :param max_loop_unroll: Maximum statically-known loop trip count to fully unroll
     :return: QIR corresponding to the HUGR input as bytes
     :rtype: bytes
     """
@@ -21,10 +28,19 @@ def guppy_to_qir_bytes(entrypoint: Callable, *, validate_qir: bool = True) -> by
         raise ValueError(message)
 
     assert isinstance(entrypoint, GuppyFunctionDefinition)  # noqa: S101
-    return to_qir_bytes(entrypoint.compile(), validate_qir=validate_qir)
+    return to_qir_bytes(
+        entrypoint.compile(),
+        validate_qir=validate_qir,
+        max_loop_unroll=max_loop_unroll,
+    )
 
 
-def guppy_to_qir_str(entrypoint: Callable, *, validate_qir: bool = True) -> str:
+def guppy_to_qir_str(
+    entrypoint: Callable,
+    *,
+    validate_qir: bool = True,
+    max_loop_unroll: int = max_loop_unroll_default(),
+) -> str:
     """
     Converts guppy entrypoint to qir str
 
@@ -32,6 +48,7 @@ def guppy_to_qir_str(entrypoint: Callable, *, validate_qir: bool = True) -> str:
     :type entrypoint: GuppyFunctionDefinition
     :param validate_qir: Whether to validate the created QIR
     :type validate_qir: bool
+    :param max_loop_unroll: Maximum statically-known loop trip count to fully unroll
     :return: QIR corresponding to the HUGR input as a str
     :rtype: str
     """
@@ -42,4 +59,8 @@ def guppy_to_qir_str(entrypoint: Callable, *, validate_qir: bool = True) -> str:
 
     assert isinstance(entrypoint, GuppyFunctionDefinition)  # noqa: S101
 
-    return to_qir_str(entrypoint.compile(), validate_qir=validate_qir)
+    return to_qir_str(
+        entrypoint.compile(),
+        validate_qir=validate_qir,
+        max_loop_unroll=max_loop_unroll,
+    )